発症日別による陽性者数の推移: Rename components/cards/PositiveNumberByDevelopedDateCard.vue to components/index/CardsReference/PositiveNumberByDevelopedDate/Card.vue
