+++ libraries/repositories/PositiveByDevelopedRepository.ts
@@ -19,7 +19,7 @@ export class PositiveByDevelopedRepository
   /**
    * 使用箇所
    *
-   * 発症日別による陽性者数の推移 (components/cards/PositiveNumberByDevelopedDateCard.vue)
+   * 発症日別による陽性者数の推移 (components/index/CardsReference/PositiveNumberByDevelopedDate/Card.vue)
    */
   constructor() {
     super(Convert.toPositiveByDeveloped(JSON.stringify(rawData)))
